journal: fixed int overflow on >2GiB changesets

+++ src/knot/journal/journal.c
@@ -750,7 +750,7 @@ static int vals_to_changeset(knot_db_val_t *vals, int nvals,
 		return KNOT_ENOMEM;
 	}
 
-	for (int i = 0; i < nvals; i++) {
+	for (size_t i = 0; i < nvals; i++) {
 		valps[i] = vals[i].data + JOURNAL_HEADER_SIZE;
 		vallens[i] = vals[i].len - JOURNAL_HEADER_SIZE;
 	}
@@ -1279,7 +1279,7 @@ static int store_changesets(journal_t *j, list_t *changesets)
 			txn->ret = KNOT_ENOMEM;
 			break;
 		}
-		for (int i = 0; i < maxchunks; i++) {
+		for (size_t i = 0; i < maxchunks; i++) {
 			chunkptrs[i] = allchunks + i*CHUNK_MAX + JOURNAL_HEADER_SIZE;
 		}
 		txn->ret = changeset_serialize(ch, chunkptrs, CHUNK_MAX - JOURNAL_HEADER_SIZE,
@@ -1291,14 +1291,14 @@ static int store_changesets(journal_t *j, list_t *changesets)
 		uint32_t serial = knot_soa_serial(&ch->soa_from->rrs);
 		uint32_t serial_to = knot_soa_serial(&ch->soa_to->rrs);
 
-		for (int i = 0; i < chunks; i++) {
+		for (size_t i = 0; i < chunks; i++) {
 			vals[i].data = allchunks + i*CHUNK_MAX;
 			vals[i].len = JOURNAL_HEADER_SIZE + chunksizes[i];
 			make_header(vals + i, serial_to, chunks);
 		}
 
 	// PART 6: inserting vals into db
-		for (int i = 0; i < chunks; i++) {
+		for (size_t i = 0; i < chunks; i++) {
 			if (txn->ret != KNOT_EOK) break;
 			txn_key_2u32(txn, j->zone, serial, i);
 			txn->val = vals[i];
